chore: refactoring oppu
