--- src/main/java/com/ppu/ppu/oppu/OppuRepository.java
@@ -1,13 +1,0 @@
-package com.ppu.ppu.oppu;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface OppuRepository extends JpaRepository<Oppu, UUID> {
-
-}
