+++ src/main/java/com/ppu/ppu/oppu/OppuService.java
@@ -3,6 +3,10 @@ package com.ppu.ppu.oppu;
 import com.ppu.ppu.exception.ErrorCode;
 import com.ppu.ppu.exception.domain.OppuException;
 import com.ppu.ppu.exception.domain.UserException;
+import com.ppu.ppu.oppu.domain.Oppu;
+import com.ppu.ppu.oppu.domain.OppuRepository;
+import com.ppu.ppu.oppu.dto.OppuGetArticleDto;
+import com.ppu.ppu.oppu.dto.OppuPostArticleDto;
 import com.ppu.ppu.user.UserRepository;
 import com.ppu.ppu.utils.image.ImageService;
 import lombok.AllArgsConstructor;
