+++ src/main/java/com/ppu/ppu/oppu/domain/OppuRepository.java
@@ -1,0 +1,9 @@
+package com.ppu.ppu.oppu.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface OppuRepository extends JpaRepository<Oppu, UUID> {
+
+}
